+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/dataEntry.epics.js
@@ -5,13 +5,13 @@ import { pipe } from 'capture-core-utils';
 import { map } from 'rxjs/operators';
 import { batchActions } from 'redux-batched-actions';
 import { convertFormToClient, convertClientToServer } from '../../../converters';
-import { dataEntryActionTypes, updateTei, setTeiModalError, setTeiAttributeValues } from './dataEntry.actions';
+import { dataEntryActionTypes, updateTei, setTeiModalError, setTeiValues } from './dataEntry.actions';
+import { GEOMETRY } from './helpers';
 
 const convertFn = pipe(convertFormToClient, convertClientToServer);
-const geometryType = (key) => {
-    const types = ['Point', 'None', 'Polygon'];
-    return types.find(type => key.toLowerCase().includes(type.toLowerCase()));
-};
+
+const geometryType = formValuesKey =>
+    Object.values(GEOMETRY).find((value: any) => value.FEATURETYPE === formValuesKey);
 
 const standardGeoJson = (geometry) => {
     if (!geometry) {
@@ -65,7 +65,7 @@ export const updateTeiEpic = (action$: InputObservable, store: ReduxStore) =>
                 trackedEntities: [
                     {
                         attributes: deriveAttributesFromFormValues(formServerValues),
-                        geometry: deriveGeometryFromFormValues(formServerValues),
+                        geometry: deriveGeometryFromFormValues(values),
                         trackedEntity: trackedEntityInstanceId,
                         trackedEntityType: trackedEntityTypeId,
                         orgUnit: orgUnitId,
@@ -87,8 +87,11 @@ export const updateTeiSucceededEpic = (action$: InputObservable) =>
     action$.pipe(
         ofType(dataEntryActionTypes.TEI_UPDATE_SUCCESS),
         map((action) => {
-            const attributeValues = action.meta?.serverData?.trackedEntities[0]?.attributes || [];
-            return batchActions([setTeiAttributeValues(attributeValues)]);
+            const trackedEntity = action.meta?.serverData?.trackedEntities[0] || {};
+            const { attributes = [], geometry } = trackedEntity;
+
+
+            return batchActions([setTeiValues(attributes, geometry)]);
         }),
     );