+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/hooks/index.js
@@ -9,3 +9,4 @@ export { useDataElements } from './useDataElements';
 export { useOptionSets } from './useOptionSets';
 export { useProgramTrackedEntityAttributes } from './useProgramTrackedEntityAttributes';
 export { useFormValidations } from './useFormValidations';
+export { useGeometryValues } from './useGeometryValues';