+++ src/core_modules/capture-core/metaDataStoreLoaders/programs/quickStoreOperations/storePrograms.js
@@ -7,8 +7,10 @@ const convert = (() => {
     const sort = (arr: Array<any>, sortBy: string = 'sortOrder') => {
         arr.sort((a, b) => {
             if (a[sortBy] == null) {
-                return 1;
-            } else if ((b[sortBy] == null)) {
+                return b[sortBy] == null ? 0 : 1;
+            }
+
+            if (b[sortBy] == null) {
                 return -1;
             }